refactor(cli): extract schedule command module
